% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddClass}
\alias{AddClass}
\title{Add large or small flock information}
\usage{
AddClass(full.data)
}
\arguments{
\item{full.data}{A clean (greenlight) file of observations}
}
\value{
data frame of full.data with additional column of class
}
\description{
Add large or small flock information
}
\details{
This function was added for detection estimation of flocks based on flock size.  It adds a class column to the full data that
specifies small or large flocks.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
