% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AdjustCounts}
\alias{AdjustCounts}
\title{Create index columns based on Num and Obs_Type}
\usage{
AdjustCounts(full.data)
}
\arguments{
\item{full.data}{A clean (greenlight) file containing observation history}
}
\value{
data frame of original data and 5 new index columns
}
\description{
AdjustCounts will create new columns for 5 indices (itotal, ibb, total, sing1pair2, and flock) and compute the values based on Num and Obs_Type
}
\details{
AdjustCounts will take a new observation file and augment the Num category to create what a particular observation means to an index.
Currently there are 5 indices that are created: \enumerate{
\item itotal - Indicated total.  Singles doubled, pairs doubled, opens added, flkdrake 1-4 doubled, flkdrake 5+ added.
\item ibb - Indicated breeding birds.  Singles doubled, pairs doubled, opens removed, flkdrake 1-4 doubled, flkdrake 5+ removed.
\item total - Total birds.  Singles added, pairs doubled, opens added, flkdrake added.
\item sing1pair2 - Singles and pairs.  Singles added, pairs doubled, opens removed, flkdrake 1-4 added.
\item flock - Flocks.  Singles removed, pairs removed, opens added, flkdrake 5+ added.
}
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
