% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AerialDataZip.R
\name{AerialDataZip}
\alias{AerialDataZip}
\title{Create a .zip file of the associated observation files, design transects, and design strata for a project}
\usage{
AerialDataZip(area = "none", drive = "K:", output.folder = "none", year)
}
\arguments{
\item{area}{The abbreviation for the project.}

\item{drive}{The drive letter referring to the location of the Waterfowl directory (for example, "K:" or "Q:").}

\item{output.folder}{The folder for the output .zip file.}
}
\value{
None, but a .zip output file is written to the output.folder or working directory.
}
\description{
AerialDataZip will combine the observation files, design transects, and design strata for a range of years for a survey and
 combine them into a .zip file created (by default) into the current working directory.
}
\details{
AerialDataZip is designed to quickly combine "official" data files into a .zip file for distribution.  It will
draw from the MasterFileList and output to the current working directory, unless another path is specified.
}
\examples{
 AerialDataZip(area="YKD", drive="T:", year=c(2000:2019))

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
