% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFrame.R
\name{AnalysisFrame}
\alias{AnalysisFrame}
\title{Filter the data objects down into an object to pass to TidyDensities}
\usage{
AnalysisFrame(survey, year)
}
\arguments{
\item{year}{The year to run.}

\item{area}{The abbreviation for the project.}
}
\value{
A list of 3 estimates tables matching ACPHistoric, YKGHistoric, YKDHistoric, or CRDHistoric
}
\description{
AnalysisFrame will compile all data for a survey-year combination in the format TidyDensities uses
}
\details{
AnalysisFrame is meant to make extracting a clean data object for TidyDensities easier
without the need for external files.  The MasterObs, MasterTransect, MasterStrata, MasterSummary, and MasterEffort
data frames must be up to date.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
