% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AppendYear}
\alias{AppendYear}
\title{Update one of the 5 tables of historic estimates (.rda objects ACPHistoric, CRDHistoric, YKDHistoric, YKGHistoric, WBPHSHistoric)}
\usage{
AppendYear(area, year)
}
\arguments{
\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKG - Yukon Kuskokwim Delta, geese
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item YKDV - Yukon Kuskokwim Delta, ducks, visibility study strata
}}

\item{year}{The year of estimates to add}
}
\value{
The associated object is updated and saved in the data folder of the package.
}
\description{
Update one of the historic estimates tables with one year of new data
}
\details{
This function updates the historic tables (available as package data .rda objects) with one year of new data.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
