% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignStrata.R
\name{AssignStrata}
\alias{AssignStrata}
\title{Assign strata ID to data for analysis using sf package}
\usage{
AssignStrata(full.data, strata.file, strata.id, retain = "liberal", area)
}
\arguments{
\item{full.data}{The data set of georeferenced observations}

\item{strata.file}{The path to the analysis stratification layer}

\item{strata.id}{The identifier of the stratification layer}

\item{area}{The survey being processed}
}
\value{
data frame of observations with modified Strata column
}
\description{
AssignStrata will provide the strata layer associated with a point in a clean data set
}
\details{
AssignStrata will calculate the nearest polygon to observations in a data file and overwrite the
Strata column in the data with the nearest polygon's strata.id
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
