% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBFcreate.R
\name{BBFcreate}
\alias{BBFcreate}
\title{Create the big beautiful file for AK Waterfowl data}
\usage{
BBFcreate()
}
\value{
The associated object is updated and saved in the data folder of the package.
}
\description{
Create the big beautiful file for AK Waterfowl data and save it to a new .rda
}
\details{
This function updates the big beautiful file with new data found in MasterObs, MasterStrata, and MasterEffort.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
