% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedReport.R
\name{BoundingByStrata}
\alias{BoundingByStrata}
\title{Calculate bounding boxes for each distinct stratum in a stratification .shp file}
\usage{
BoundingByStrata(strata.path)
}
\arguments{
\item{strata.path}{The path to the stratification .shp file.}
}
\value{
List of distinct strata and their bounding boxes.
}
\description{
BoundingByStrata calculates a bounding box for each stratum in a .shp file in a format leaflet can use to display in markdown efficiently
}
\details{
BoundingByStrata will calculate separate bounding boxes for each stratum in a .shp file.  It is primarily used to display each
stratum specifically in leaflet used within Rmarkdown.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
