% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{ColSwitch}
\alias{ColSwitch}
\title{Switch necessary column names}
\usage{
ColSwitch(data)
}
\arguments{
\item{data}{The data frame to be checked.}
}
\value{
Data frame with new column names.
}
\description{
ColSwitch takes old column names and changes them to current styling
}
\details{
ColSwitch will change yr, mo, da, se, strata, grp, and sppn to Year, Month, Day, Seat, Strata, Obs_Type, and Species, respectively.
This function was written to quickly allow an old data file to be formatted for analysis, but probably won't be used in the future.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
