% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedReport.R
\name{CombinedReport}
\alias{CombinedReport}
\title{Runs a new combination of separate and combined estimates for 2 observers}
\usage{
CombinedReport(
  strata.path = NA,
  transect.path = NA,
  data1.path = NA,
  data2.path = NA,
  area = NA,
  threshold = 0.5
)
}
\arguments{
\item{strata.path}{The file location of the analysis stratification .shp file}

\item{transect.path}{The file location of the transect design .shp file}

\item{data1.path}{The file location of the .csv transcribed data for observer 1}

\item{data2.path}{The file location of the .csv transcribed data for observer 2}

\item{area}{The area code for dedicated MBM Alaska region surveys.  Either folder.path or area must be specified.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, for either geese or ducks
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}

\item{threshold}{The distance in kilometers from a design file where observations are no longer counted.  Defaults to 0.5 km.}
}
\value{
None
}
\description{
CombinedReport will produce an annual estimate report in .html format for 2 observers
}
\details{
CombinedReport is designed to produce a basic .html report for quality checks for a particular year/crew combination. The report
contains basic visual QA/QC maps of stratification and transect design, as well as observation information by observer. It also
presents the basic and expanded table of estimates.  It uses the markdown script within AKaerial (CombinedSurveyReport.Rmd) and writes an
output report to the directory of data1.path.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
