% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CombineEstimates}
\alias{CombineEstimates}
\title{Combine estimates and variances for an observer crew}
\usage{
CombineEstimates(estimates)
}
\arguments{
\item{estimates}{The \code{estimates} element of the list returned in Densities}
}
\value{
data frame of pooled estimates
}
\description{
CombineEstimates will take observer-specific estimates and summarize them as a crew estimate for a given year
}
\details{
CombineEstimates will take observer-specific estimates and summarize them as a crew estimate for a given year
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
