% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{CombineEstimatesByStrata}
\alias{CombineEstimatesByStrata}
\title{Combine WBPHS estimates by stratum}
\usage{
CombineEstimatesByStrata(estimates)
}
\arguments{
\item{estimates}{WBPHS index estimates from EstimatesWBPHS.}
}
\value{
Data frame of combined index estimates by species at the stratum level
}
\description{
CombineEstimatesByStrata will combine multiple observer estimates at the stratum level
}
\details{
CombineEstimatesByStrata takes an estimate object (See EstimatesWBPHS) and merges the estimates of 2 observers by species at the stratum level.
The resulting data frame is appended onto the master WBPHS results table.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
