% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CorrectionFactor}
\alias{CorrectionFactor}
\title{Apply a detection correction to an index estimate}
\usage{
CorrectionFactor(estimates, species)
}
\arguments{
\item{estimates}{The estimates list object from a Densities output list}

\item{species}{The character string or vector representing the species to be corrected.}
}
\value{
original estimates file with additional detection-corrected column
}
\description{
CorrectionFactor will apply a detection correct to an index estimate.
}
\details{
CorrectionFactor currently only works for dusky Canada geese (DCGO).
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
