% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CorrectTrans}
\alias{CorrectTrans}
\title{Replace navigational transect (tran) with corrected transect (ctran)}
\usage{
CorrectTrans(full.data, threshold = 0.5, split.design, area, strata.file)
}
\arguments{
\item{full.data}{A clean (greenlight) file of observations}

\item{threshold}{The distance in kilometers from a design file where observations are no longer counted.  Defaults to 0.5 km.}

\item{split.design}{SplitDesign output object of corrected transect lines}

\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
\item YKG - Yukon Kuskokwim Delta, geese
\item KIG - Kigigak Island
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}

\item{strata.file}{The path to the .shp file of the stratification}
}
\value{
original data frame (full.data) with ctran column added for corrected transect number
}
\description{
CorrectTrans will cross-reference original navigational numbering of transects with corrected SplitDesign numbering
}
\details{
CorrectTrans uses SplitDesign output to overwrite potentially incorrect navigational transect numbering with sample-appropriate numbering.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
