% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CountsTable}
\alias{CountsTable}
\title{Summarize indices by year, observer, transect, species, and strata}
\usage{
CountsTable(adj.counts)
}
\arguments{
\item{adj.counts}{A data frame that has been through AdjustCounts}
}
\value{
data frame with tabular summary of observations
}
\description{
CountsTable will summarize an AdjustCounts data frame into a table
}
\details{
CountsTable is designed to be run on an adjusted data file (has been run through AdjustCounts to create indices).  It will return a tabular summary of counts.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
