% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{CountsWBPHS}
\alias{CountsWBPHS}
\title{Summarize index counts for WBPHS data by Transect and Stratum}
\usage{
CountsWBPHS(adj.counts)
}
\arguments{
\item{adj.counts}{An R data frame of observations in a given year of the survey}
}
\value{
A data frame with index values by Year, Observer, Species, Transect, and Stratum.
}
\description{
Summarize index counts for WBPHS data by Transect and Stratum
}
\details{
This function takes standard greenlight data for the WBPHS (Waterfowl Breeding Population Habitat Survey, or North American) and
summarizes it at the Transect and Stratum levels. It uses already adjusted counts by Obs_Type (see AdjustCounts).
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
