% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{Cut9}
\alias{Cut9}
\title{Clip stratum 9 for overlap for 3 species}
\usage{
Cut9(full.data)
}
\arguments{
\item{full.data}{An R data frame of observations in a given year of the survey}
}
\value{
The same data frame with modified Stratum 9 observations.
}
\description{
Clip the data from stratum 9 for CCGO, GWFG, and SWAN observations past a longitudinal gradient
}
\details{
This function removes observations in Stratum 9 by a pre-defined longitudinal gradient for swans,
cackling Canada geese, and greater white-fronted geese.  These observations are removed prior to index estimation since they are included in
other MBM coastal zone surveys (YKG).  The gradient is as follows, by transect number:
 \itemize{
  \item 12:  -163.559
  \item 13:  -163.738
  \item 14:  -164.388
  \item 15:  -164.130
  \item 16:  -164.440
  \item 17:  -164.995
  \item 18:  -164.938
  \item 19:  -164.810
  }
Observations for CCGO, GWFG, and SWAN are trimmed past the western edge of the gradient and remaining observations are now attributed to
stratum 99 for documentation and to avoid confusion with stratum 9.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
