% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DataSelect}
\alias{DataSelect}
\title{Format and summarize a clean data set for analysis}
\usage{
DataSelect(
  area,
  data.path = NA,
  strata.path = NA,
  transect.path = NA,
  threshold = 0.5
)
}
\arguments{
\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
\item YKG - Yukon Kuskokwim Delta, geese
\item KIG - Kigigak Island
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}

\item{data.path}{The location of the QA/QC checked data file}

\item{strata.path}{The location of the analysis stratification .shp file}

\item{transect.path}{The location of the transect design .shp file}

\item{threshold}{The distance in kilometers from a design file where observations are no longer counted.  Defaults to 0.5 km.}
}
\value{
List object with 4 elements: \enumerate{
\item obs The original data file with corrected transect and stratification information
\item flight The flown transects for the observer with corrected numbering
\item design The split design information
\item strata The spatial polygon summary for the stratification layer
}
}
\description{
DataSelect will combine observation, design transect, and stratification layers into a summary list for analysis
}
\details{
DataSelect reads in a clean data file (passed through GreenLight function) and pulls all observations with code==1.  All observations
coded as open 1 are then changed to single 1.  The transect layer is then placed on the stratification and trimmed and renumbered from
navigational numbers to analysis numbering.  Species codes that need specific project treatments will be changed.  Counts are adjusted
for different indices at the row level. The resulting list object is the primary currency for design-based ratio estimator analysis.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
