% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Input number in degrees}
}
\value{
Radian conversion
}
\description{
Convert degrees to radians.
}
\details{
Convert degrees to radians
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
