% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceToDesign.R
\name{DistanceToDesign}
\alias{DistanceToDesign}
\title{Distance to the reported design transect}
\usage{
DistanceToDesign(full.data, design.trans, trans.id)
}
\arguments{
\item{full.data}{The data set of georeferenced observations}

\item{design.trans}{An sf object of the design transects}

\item{trans.id}{The transect column reflecting the design numbering}

\item{transect.path}{The acceptable distance before removal of the point in km}
}
\value{
data frame of observations with modified Distance column
}
\description{
DistanceToDesign will provide the shortest distance to the reported design transect associated with a point in a clean data set
}
\details{
DistanceToDesign will calculate the distance to the nearest design transect
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
