% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DubObs.R
\name{DoubleObsRandom}
\alias{DoubleObsRandom}
\title{Derive a random double observer sampling design}
\usage{
DoubleObsRandom(strata, transects, split = 0.4, options = 5, tries = 10000)
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{solution}{The solution from a \code{\link{DoubleObsRandom}} optimization.}
}
\value{
None
}
\description{
DoubleObsRandom will load an existing stratified transect design and assign a third
observer approximately equally by strata and area randomly behind the pilot or observer
}
\details{
DoubleObsRandom is designed to take 2 spatial objects: a stratification base layer
and transect layer represented as spatial lines, and assign a third observer (during
a double-observer detection study) approximately equally behind both front seat observers.
The function simulates possible random combinations of left and right seat and will throw
out those that aren't within \code{split} proportion.  The resulting map and design will
offer n options (\code{options} defaults to 5) that can then be checked by the field crew
for logistical feasibility.
}
\examples{
 ShowMeDouble(strata.path="C:/Habitat.shp", solution=DoubleObsRandom(strata=MyStrata, transects=MyTransects))

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
