% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DubObs.R
\name{DubMatch}
\alias{DubMatch}
\title{Match observations for double observer surveys}
\usage{
DubMatch(
  front.data = NA,
  rear.data = NA,
  combined.data = NA,
  front.obs,
  rear.obs,
  time = 6,
  open = 5
)
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{solution}{The solution from a \code{\link{DoubleObsRandom}} optimization.}
}
\value{
Returns a leaflet object for display.
}
\description{
DubMatch attempts to match observations between 2 observers on a survey by time and group size.
}
\details{
ShowMeDouble is designed to take a directory path for a stratification base layer
and the result of \code{\link{DoubleObsRandom}} and produce an interactive Leaflet map.
Clicking on strata will bring up the STRATNAME field associated with the click.
Clicking on a transect will display the associated strata identification, transect numbering,
recalculated (SPLIT) transect number (see \code{\link{SplitDesign}}), and
optimal positioning of the third observer (behind pilot or front-seat observer).
}
\examples{
 ShowMeDouble(strata.path="C:/Habitat.shp", solution=DoubleObsRandom(strata=MyStrata, transects=MyTransects))

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
