% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{EstimatesTable}
\alias{EstimatesTable}
\title{Create a master table of estimates for an area}
\usage{
EstimatesTable(area, year)
}
\arguments{
\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
\item YKG - Yukon Kuskokwim Delta, geese
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
}}

\item{year}{The range of years to provide estimates over}
}
\value{
List object with 3 elements: \enumerate{
\item output.table Observer-specific estimates
\item expanded.table Deeper level index information by transect, strata, species, and observer
\item combined Combined crew index estimates by year
}
}
\description{
EstimatesTable will calculate index estimates over a range of years for a given area
}
\details{
EstimatesTable is coded to use a single observer some years and pool more than 2 observers in other years.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
