% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{EstimatesTableWBPHS}
\alias{EstimatesTableWBPHS}
\title{Deprecated function to calculate index estimates for a given year of WBPHS data}
\usage{
EstimatesTableWBPHS(year)
}
\arguments{
\item{year}{Four digit year of the WBPHS survey to use}
}
\value{
List object with 2 elements: \enumerate{
\item output.table Observer-specific estimates for the species indicated in the sppntable estimates column
\item expanded.table Deeper level count information by transect, strata, species, and observer
}
}
\description{
EstimatesTableWBPHS will load a given year of greenlighted data and calculate an index estimate using a ratio estimator.  This function is deprecated and replaced by WBPHStidy.
}
\details{
EstimatesTableWBPHS provides a shortcut to running piecewise functions for creating an annual index estimate for the WBPHS survey.
The function uses MasterFileList_WBPHS as a lookup table to locate the appropriate raw observation files, then calls ReadWBPHS to concatenate,
SummaryWBPHS for spatial design summary, TransDataWBPHS to table observations appropriately, and finally EstimatesWBPHS to calculate the ratio
estimate.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
