% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{EstimatesWBPHS}
\alias{EstimatesWBPHS}
\title{Standard ratio estimator for WBPHS survey data}
\usage{
EstimatesWBPHS(data, flight)
}
\arguments{
\item{data}{The ReadWBPHS list object to be analyzed}

\item{flight}{Flight summary from SummaryWBPHS}
}
\value{
List object with 2 elements: \enumerate{
\item output.table Observer-specific estimates for the species indicated in the sppntable estimates column
\item expanded.table Deeper level count information by transect, strata, species, and observer
}
}
\description{
EstimatesWBPHS will combine spatially-referenced observations with design transects and strata to create an index estimate
}
\details{
EstimatesWBPHS is the primary function in AKaerial for producing index estimates for the WBPHS survey. It will take an object from ReadWBPHS and adjust counts, summarize
spatial information, and calculate indices and their associated standard errors.  Also retained are estimates of densities of bird by strata and on
 each transect.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
