% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FindVoids}
\alias{FindVoids}
\title{Find gaps between polygons of the same stratum type}
\usage{
FindVoids(pieces)
}
\arguments{
\item{pieces}{Polygons within a given stratum}
}
\value{
data frame of voids and lengths
}
\description{
FindVoids will find the gaps between discontinuous polygons and remove them from potential available transect calculations
}
\details{
Findvoids is used in the maximum available transect (M) calculation in the variance estimate in the ratio estimator. It removes gaps
in strata from the maximum-minimum latitude calculation.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
