% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinishEstimate.R
\name{FinishEstimate}
\alias{FinishEstimate}
\title{Perform final filters and additions to an estimates object by survey}
\usage{
FinishEstimate(output.table, expanded.table, combined)
}
\value{
A clean estimates object
}
\description{
FinishEstimate will filter and add as appropriate to an estimates object
}
\details{
FinishEstimate checks the appropriate species list for the final estimates object and will add
or remove entries as appropriate. This includes START and END.  It will also combine observer estimates
into one combined table where appropriate.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
