% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FixTavs}
\alias{FixTavs}
\title{FixTavs runs near the end of DataSelect to overwrite species codes from CCGO to TAVS.}
\usage{
FixTavs(selected.data)
}
\arguments{
\item{selected.data}{A nearly complete DataSelect object from the Yukon Kuskokwim Delta}
}
\value{
original data frame (full.data) with some CCGO species codes changed to TAVS
}
\description{
FixTavs runs near the end of DataSelect to overwrite species codes from CCGO to TAVS.
}
\details{
FixTavs runs near the end of DataSelect to overwrite species codes from CCGO to TAVS.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
