% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullHistoric.R
\name{FullHistoric}
\alias{FullHistoric}
\title{Recreate the entire ACPHistoric, YKDHistoric, YKGHistoric, and CRDHistoric estimates objects}
\usage{
FullHistoric()
}
\value{
4 lists of 3 estimates tables matching ACPHistoric, YKGHistoric, YKDHistoric, and CRDHistoric
}
\description{
FullHistoric will re-run the entire estimates process for all years
}
\details{
FullHistoric will take the current data objects providing strata, transect, and observation
information and run the entire historic estimates process over all years for ACP, YKD, YKG, and CRD.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
