% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherData.R
\name{GatherData}
\alias{GatherData}
\title{Gather all data for a project and add unique design panel information}
\usage{
GatherData(area = "none", drive = "K:", ouput.folder = "none")
}
\arguments{
\item{area}{The abbreviation for the project.}

\item{drive}{The drive letter referring to the location of the Waterfowl directory (for example, "K:" or "Q:").}

\item{output.folder}{The folder for 3 output files: flight.csv, obs.csv, and transect.csv, defaults to the working directory.}
}
\value{
None, but 3 output files are written to the output.folder or working directory.
}
\description{
GatherData will compile all data for a project over its history, add transect panel information (if available), and a unique transect identifier
}
\details{
GatherData is designed to compile all data over a range of years for a project with a given design transect layer and design
stratification.  It will add a series of columns to the resulting data that are used to derive the unique identifier for a transect
that may have also occurred in other years (for a panel design).
}
\examples{
 GatherData(area="YKD", drive="T:")

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
