% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gcd.slc}
\alias{gcd.slc}
\title{Spherical distance between 2 points}
\usage{
gcd.slc(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{Longitude of point 1 in degrees}

\item{lat1}{Latitude of point 1 in degrees}

\item{long2}{Longitude of point 2 in degrees}

\item{lat2}{Latitude of point 2 in degrees}
}
\value{
distance in km
}
\description{
Calculates the geodesic distance between two points specified by radian latitude/longitude using the Spherical Law of Cosines (slc)
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
