% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LoadMap}
\alias{LoadMap}
\title{Read .shp file and attach strata names to the object}
\usage{
LoadMap(map.file, type = "df")
}
\arguments{
\item{map.file}{The path to the .shp file to be read in}

\item{type}{Either "df" to return a data frame or "proj" to return a WGS84 projected spatial object}
}
\value{
data frame or projected spatial object
}
\description{
LoadMap will read in a .shp file and return either a data frame or projected spatial object
}
\details{
LoadMap is designed to read in a .shp file and attach the strata names to a data frame and return it.  It can also return a projected
WGS84 surface.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
