% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MakeStateSpace}
\alias{MakeStateSpace}
\title{Iterate over all species and projects and create input files for use in StateSpace}
\usage{
MakeStateSpace(folder.path, area = "all")
}
\arguments{
\item{folder.path}{The output folder to send newly generated files to}
}
\value{
Many .csv files are created in folder.path
}
\description{
Iterate over all species and projects and create input files for use in StateSpace
}
\details{
Iterate over all species and projects and create input files for use in StateSpace.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
