% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2line.R
\name{Mappoints2line}
\alias{Mappoints2line}
\title{Display a linestring object from a series of transect observations}
\usage{
Mappoints2line(area, year)
}
\arguments{
\item{obs}{sf object of bird observations with referenced Lat Lon}

\item{Year}{year of the observations}

\item{Transect}{transect number the observations were associated with}
}
\value{
sf Linestring object
}
\description{
Mappoints2line displays a points2line Linestring object.
}
\details{
This function will connect observations into a linestring object to attempt to recreate minimum sampling effort,
then display it.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Erik Osnas, \email{erik_osnas@fws.gov}
}
