% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassReport.R
\name{MassPool}
\alias{MassPool}
\title{Pool data from multiple observers across all MBM surveys where appropriate}
\usage{
MassPool()
}
\description{
MassPool will selectively apply the \code{\link{PoolData}} function to all known survey, year, and observer
combinations that need pooling.
}
\details{
Pooling data files becomes necessary when observers and pilots switch seats or
are replaced during a survey.  The resulting files are generally pooled by seat side (left or right).
MassPool currently applies the \code{\link{PoolData}} function to the following data files:
\itemize{
\item 1986 YKG files from observers C. Lensink and K. Bollinger are pooled into a right front seat file.
Lensink observed from 6-2-1986 through 6-22-1986 with the exception of 6-13-1986.  Bollinger observed on 7 design transects on 6-13-1986.
W. Butler was the only pilot for the 1986 YKG survey.
\item 1988 CRD files
}
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
