% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassReport.R
\name{MassReport}
\alias{MassReport}
\title{Apply \code{\link{GreenLight}} function to an entire folder}
\usage{
MassReport(folder.path = NA, area = NA, method = "both")
}
\arguments{
\item{folder.path}{The path name to the folder containing the raw .csv files.
Either folder.path or area must be specified.}

\item{area}{The area code for dedicated MBM Alaska region surveys.  Either folder.path or area must be specified.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, for either geese or ducks
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
}}

\item{method}{Specify \code{greenlight}, \code{data}, or \code{both}.
\itemize{
\item \code{method=greenlight} will run the standard \code{\link{GreenLight}} report for each data file
\item \code{method=data} will generate the QAQC data file (if possible) for each data file
\item \code{method=both} will run the standard \code{\link{GreenLight}} report for each data file and produce each associated QAQC data file
}}
}
\value{
None
}
\description{
MassReport will scan a folder for .csv data files and apply the \code{\link{GreenLight}}
function to each.
}
\details{
MassReport is designed to produce QAQC reports for a large number of files based on a set of established criteria for data integrity.
It can also create the files used for analysis (those with extraneous species codes removed) based on the method chosen.
It was created to avoid single calls to Greenlight when changes are implemented in the QAQC process that need to be applied to all data files at once, such as
including or excluding a species in the generated sharable analysis files.
}
\examples{
 MassReport(area="ACP", method="both")

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
