% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{ObsByYear}
\alias{ObsByYear}
\title{Split an old combined data file by observer and year}
\usage{
ObsByYear(path.name)
}
\arguments{
\item{path.name}{The path to the combined file.}
}
\value{
None
}
\description{
ObsByYear will take a .csv or .txt file with 2 or more observers and parse it into observer- and year-specific files
}
\details{
ObsByYear will take a .csv or .txt file with 2 or more observers and parse it into observer- and year-specific files. This is designed to fix
older files that may have been concatenated into a master file, splitting them into separate pieces.  Since this was an ACP problem,
the files default to the directory "Q:/Waterfowl/Parsed/ACP_" and would need recoding for any other survey.  Individual .csv files are written to the directory.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
