% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{ObserverCheck}
\alias{ObserverCheck}
\title{Check if observer initials are uppercase.}
\usage{
ObserverCheck(data)
}
\arguments{
\item{data}{The data frame to be checked.}
}
\value{
TRUE/FALSE
}
\description{
ObserverCheck checks to see if the observer initials were entered consistently in uppercase
}
\details{
ObserverCheck checks to see if the observer initials were entered consistently in uppercase.  This is a basic QA/QC check
so the analysis is not split among upper and lowercase versions of the same observer.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
