% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsMap.R
\name{ObsMap}
\alias{ObsMap}
\title{Display a combination of strata, transect design, and data}
\usage{
ObsMap(
  area = "none",
  strata,
  transect,
  data,
  year = 0,
  observer = "both",
  species = "all",
  type = "static",
  mirror = "NULL"
)
}
\arguments{
\item{area}{The project (ACP, CRD, YKD, YKG) designation, takes the place of strata and transect,
and when combined with year, also chooses the transect layer and data files.}

\item{strata}{The path name to the .gpkg file of the stratification.}

\item{transect}{The path name to the .gpkg file of lines designating the transect design.}

\item{data}{The path name(s) to the data file(s).}

\item{year}{The year to be displayed.}

\item{observer}{The observer initials to be displayed.}

\item{species}{The species code(s) to be displayed.}

\item{type}{Static (default) or dynamic map.}

\item{mirror}{Optional mirror location for offline observation and spatial data}
}
\value{
Map object
}
\description{
ObsMap will take input layers of strata, transect design, and data and display them.
}
\details{
ObsMap is designed to take a directory path for a combination of stratification base layer,
aerial transect design, and resulting data and display it using an interactive map.
}
\examples{
 ObsMap(area="CRD", year=2024, species="DCGO")

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
