% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhotoEstimate.R
\name{PhotoEstimate}
\alias{PhotoEstimate}
\title{Produce replicate-specific photo count estimates.}
\usage{
PhotoEstimate(data, total.area, M)
}
\arguments{
\item{data}{An R data frame of photo counts and areas.  Needs a minimum structure with columns rep, transect, count, and area.}

\item{total.area}{The size of the study area to extrapolate density over.}

\item{M}{The maximum possible transects through the study area (used in variance calculation)}
}
\value{
A list of 2 objects: rep_summary for information regarding the variance, and est for the estimate by replicate.
}
\description{
Produce replicate-specific photo count estimates for transect-based photo count surveys.
}
\details{
This function takes a data set of photos and associated automated or manual counts and
produces an estimate across transects at the replicate level.
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
