% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChanges.R
\name{PlotChanges}
\alias{PlotChanges}
\title{Plot design strata estimates vs analysis strata estimates}
\usage{
PlotChanges(survey, species, index = "total")
}
\value{
A ggplot object of design strata vs analysis strata estimates
}
\description{
PlotChanges will take the design estimates and analysis strata estimates and plot them
}
\details{
PlotChanges will take the design estimates and analysis strata estimates and plot them
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
