% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2line.R
\name{points2line}
\alias{points2line}
\title{Create a Linestring object from a series of transect observations}
\usage{
points2line(
  obs,
  Year = unique(obs$Year),
  Transect = unique(obs$Transect),
  crs = 4326
)
}
\arguments{
\item{obs}{sf object of bird observations with referenced Lat Lon}

\item{Year}{year of the observations}

\item{Transect}{transect number the observations were associated with}
}
\value{
sf Linestring object
}
\description{
points2line connects georeferenced points into an sf Linestring object.
}
\details{
This function will connect observations into a linestring object to attempt to recreate minimum sampling effort.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Erik Osnas, \email{erik_osnas@fws.gov}
}
