% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PoolData}
\alias{PoolData}
\title{Combine multiple observers' transcribed files in a given year}
\usage{
PoolData(year, area)
}
\arguments{
\item{year}{The year to pool}

\item{area}{The area code for dedicated MBM Alaska region surveys.
   Acceptable values include:
 \itemize{
 \item YKD - Yukon Kuskokwim Delta, ducks
 \item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
 \item YKG - Yukon Kuskokwim Delta, geese
 \item ACP - Arctic Coastal Plain
 \item CRD - Copper River Delta
 }

 @return A .csv file is written to the data directory of pooled observations}
}
\description{
Combine multiple observers' transcribed files in a given year
}
\details{
PoolData uses MasterFileList to iterate through and appropriately pool observers in certain years
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
