% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popest.R
\name{popest}
\alias{popest}
\title{Title}
\usage{
popest(dat = NA, vcfData = NA, Mdat = NA, prob = FALSE)
}
\arguments{
\item{dat}{A data frame containing transect-level sums of observed singles and pairs, see details.}

\item{vcfData}{A data frame containing visibility correction factors by strata, see details.}

\item{prob}{A logical indicating where to do the calculation based on detection probability or it reciprocal, vcf.}

\item{M}{A named vector giving the total number of possible transects in each strata. Names must match strata names in data.}
}
\value{
A list containing population estimates of 'indicated breeding birds', the 'index" estimates (without detection correction), standard errors of each, and a ggplot objects of each by year.
}
\description{
Title
}
