% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popestACP.R
\name{popestACP}
\alias{popestACP}
\title{A function to calculate variance of ACP population estimates}
\usage{
popestACP(dat = NA, pData = NA, Mdat = NA)
}
\arguments{
\item{dat}{A data frame containing transect-level sums of observed singles and pairs, see details.}

\item{pData}{A data frame containing visibility correction factors by strata, see details.}

\item{Mdat}{A data frame ... giving the total number of possible transects in each strata. Names must match strata names in data.}

\item{prob}{A logical indicating where to do the calculation based on detection probability or it reciprocal, vcf.}
}
\value{
A list containing population estimate of 'indicated breeding birds', the 'index" estimates (without detection correction), standard errors of each, and a ggplot objects of each by year.
}
\description{
A function to calculate variance of ACP population estimates
}
