% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessEstimate.R
\name{ProcessEstimate}
\alias{ProcessEstimate}
\title{Process a raw estimate into a useful table}
\usage{
ProcessEstimate(data)
}
\arguments{
\item{data}{AnalysisFrame output object consisting of summary information}
}
\value{
A list of combined, expanded.table, and output.table data frames
}
\description{
ProcessEstimate will take an AnalysisFrame object and summarize it into estimates
}
\details{
ProcessEstimate will take an AnalysisFrame object and summarize it into estimates. The
usual output is a simple observer-level file, but it gets translated into observer-, strata-,
and combined estimates.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
