% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFiles.R
\name{random_files}
\alias{random_files}
\title{Randomly sample files in a directory}
\usage{
random_files(
  path,
  percent_number,
  pattern = "jpg$|JPG$",
  folders = TRUE,
  results = "Q:/Waterfowl/STEI_Survey/Data/RandomFiles"
)
}
\arguments{
\item{path}{The path to the folder containing files to be sampled (or folders of files).}

\item{percent_number}{The percent (if less than 1.0) or raw number (if greater than 1.0) of files to sample.}

\item{pattern}{The file extension of the requested sample.}

\item{folders}{Should nested folders also be searched?  TRUE/FALSE}

\item{results}{The folder path for the copied (sampled) output}
}
\value{
copies randomly selected files to results directory and writes a lookup key \code{selected_files.csv}
}
\description{
RandomFiles will randomly sample files from a directory and optionally its internal folders
}
\details{
RandomFiles was designed to randomly sample systematic photos of a sampled area for analysis.  It is generalized to use any file extension
a user inputs.  Files are copied and a lookup table is generated to crosswalk original and new names.
}
\examples{
 RandomFiles(path = "C:/DATA/PHOTOS", percent_number = 0.10, pattern = "jpg$|JPG$", folders = TRUE, results = "C:/RESULTS")

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
