% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{Raw2Analysis}
\alias{Raw2Analysis}
\title{Check if common fixes need to be applied to a raw data file}
\usage{
Raw2Analysis(data, area)
}
\arguments{
\item{data}{The data frame to be checked.}

\item{area}{The project area designation.}
}
\value{
list of data and fix categories that should be applied
}
\description{
Raw2Analysis checks to see if any of the common fixes need to be applied to a data set.
}
\details{
Raw2Analysis is a QA/QC check in GreenLight that will check for and apply common fixes to a data set through CommonFix.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
