% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ReportFigure}
\alias{ReportFigure}
\title{Create a basic figure from historic estimates}
\usage{
ReportFigure(
  area,
  species,
  year = "all",
  type = "combined",
  index,
  index2 = "none",
  title = "none",
  x.title = "Year",
  y.title = "Population Index"
)
}
\arguments{
\item{area}{Which area (project) will be plotted (ACP, CRD, YKD, YKG)}

\item{species}{The species chosen (see \code{\link{sppntable}} for options).}

\item{year}{The range of years for the figure.}

\item{type}{Observer-specific (observer), strata-specific (strata), or combined (combined; default value)}

\item{index}{The population index requested (total, itotal, ibb, sing1pair2, flock)}

\item{index2}{The second index to be plotted simultaneously (optional).}

\item{title}{The overall figure title.  If nothing entered as an argument, the user will be prompted to enter a title in the function.}

\item{x.title}{The desired x axis label for the final figure.}

\item{y.title}{The desired y axis label for the final figure.}
}
\value{
Renders and returns a ggplot figure
}
\description{
ReportFigure will summarize historic estimates from Alaska Region aerial surveys in a point graph that can be exported for reporting.
}
\details{
ReportFigure will take one of the historic estimates tables (package data for AKaerial) and display it as a ggplot figure
suitable for an external report.
}
\examples{
ReportFigure(area="ACP",
species=c("MALL", "SPEI"),
index="ibb",
index2="itotal",
title="Mallard and Spectacled Eider population indices from the Arctic Coastal Plain, 2007-2024")

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
