% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ReportTable}
\alias{ReportTable}
\title{Create a basic exportable table from historic estimates, including n-year average}
\usage{
ReportTable(
  data,
  species,
  year,
  index = "none",
  yr.avg,
  cap = "none",
  new.names = "none"
)
}
\arguments{
\item{data}{The estimates object.  Must be in a format with Year, Species, index, and index variance if not one of
the included package data objects (ACPHistoric, CRDHistoric, YKDHistoric, YKDVHistoric, or YKGHistoric).}

\item{species}{The species chosen (see \code{\link{sppntable}} for options).  Currently
supports no more than 1 species.}

\item{year}{The range of years for the table.}

\item{index}{The column names that specify the index estimate and its variance.
Must be ordered correctly (index, variance) or (index1, variance1, index2, variance2).  Currently up to 2 indices are supported.
If no index is specified, the user will be prompted with choices from the data.}

\item{yr.avg}{The number of years in the n-year running average requested.}

\item{cap}{The overall table caption.  If nothing entered as an argument, the user will be prompted to enter a caption in the function.}

\item{new.names}{The desired column names for the final table.  If these are not specified the user will be prompted to
input within the function.}
}
\value{
Renders an .html table and also returns it in data frame format
}
\description{
ReportTable will summarize historic estimates from in an .html table that can be exported for reporting.
}
\details{
ReportTable will take one of the historic estimates tables (package data for AKaerial) and display it as an .html table
object suitable for an external report.  The user can specify n for an n-year average if one is desired.  A similarly
structured data frame could be used in place of one of the historic estimates tables, provided it has columns for Year,
Species, an index, and that index variance.  Table caption and column headings can either be specified as arguments or
entered by the user once the table is created
}
\examples{
ReportTable(data=ACPHistoric$combined, species="SPEI", year = c(2007:2019), index = c("total", "total.var"), yr.avg=3, cap="Test table!", new.names=c("Year", "Total Birds", "SE", "3-year Avg", "SE"))

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
