% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlightscribe.R
\name{Scribe2Analysis}
\alias{Scribe2Analysis}
\title{Check if common fixes need to be applied to a raw data file}
\usage{
Scribe2Analysis(data, area)
}
\arguments{
\item{data}{The data frame to be checked.}

\item{area}{The project area designation.}
}
\value{
list of data and fix categories that should be applied
}
\description{
Scribe2Analysis checks to see if any of the common fixes need to be applied to a data set.
}
\details{
Scribe2Analysis is a QA/QC check in GreenlightScribe that will check for and apply common fixes to a data set through CommonFix.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
