% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlightscribe.R
\name{ScribeMerge}
\alias{ScribeMerge}
\title{Merge the daily SCRIBE files into a single project record for an observer}
\usage{
ScribeMerge(folder, string, addWBPHS = FALSE)
}
\arguments{
\item{folder}{The folder path containing the files.}

\item{string}{The character string common to the files to be combined.}

\item{addWBPHS}{TRUE/FALSE to add columns to convert HQ format to AK format}
}
\value{
No return value, but output is a new combined csv file within the folder.
}
\description{
ScribeMerge takes an input string and matches file names within a folder, then combines them into a single .csv file.
}
\details{
SCRIBE software creates daily observation records that are most useful compiled into a complete observer record for a survey session.
ScribeMerge takes an input string and matches file names within a folder, then combines them into a single .csv file.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
