% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{ShouldBeNumeric}
\alias{ShouldBeNumeric}
\title{Check if entries in a column are all numeric}
\usage{
ShouldBeNumeric(col.data)
}
\arguments{
\item{col.data}{The column to be checked.}
}
\value{
data frame of TRUE/FALSE and offending entries
}
\description{
ShouldBeNumeric checks to see if column entries are numeric values needed for analysis.
}
\details{
ShouldBeNumeric is a GreenLight QA/QC check performed on a data set.  If the column entries are not numeric values it will return a "red" status.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
