% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMe}
\alias{ShowMe}
\title{Display a combination of strata, transect design, and data}
\usage{
ShowMe(strata.path, transect.path, data.path, species = "all")
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{transect.path}{The path name to the .shp file of lines designating the transect design.
Must be in a format accepted by \code{\link{SplitDesign}}.}

\item{data.path}{The path name(s) to the data file(s).  At minimum, data file must have columns
Observer, Species, Obs_Type, Num, and Transect.}

\item{species}{The species code(s) to be displayed.
Acceptable values are those in \code{sppntable}.}
}
\value{
None
}
\description{
ShowMe will take input layers of strata, transect design, and data and display them using Leaflet.
}
\details{
ShowMe is designed to take a directory path for a combination of stratification base layer,
aerial transect design, and resulting data and display it using an interactive Leaflet map.
Data can be more than one file, and can be filtered to a single species or group of species. Clicking on strata will bring up
the STRATNAME field associated with the click.  Clicking on a transect will display the
associated strata identification, transect numbering, recalculated (SPLIT) transect number
(see \code{\link{SplitDesign}}), and length in kilometers. Clicking an observation will display
the observer initials, species code, observation type, number of the type observed, and the
transect number \emph{reported by the pilot or observer.}
}
\examples{
 ShowMe(strata.path="C:/Habitat.shp", transect.path="My2016Transects.shp", data.path="My2016obs.csv", species=c("SPEI", "STEI"))

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
