% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMeDesign}
\alias{ShowMeDesign}
\title{Display a combination of strata and transect design}
\usage{
ShowMeDesign(strata.path, transect.path)
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{transect.path}{The path name to the .shp file of lines designating the transect design.
Must be in a format accepted by \code{\link{SplitDesign}}.}
}
\value{
None
}
\description{
ShowMeDesign will take input layers of strata and transect design and display them using Leaflet.
}
\details{
ShowMeDesign is designed to take a directory path for a combination of stratification base layer,
aerial transect design and display it using an interactive Leaflet map.
Clicking on strata will bring up the STRATNAME field associated with the click.
Clicking on a transect will display the
associated strata identification, transect numbering, recalculated (SPLIT) transect number
(see \code{\link{SplitDesign}}), and length in kilometers.  Similar to \code{\link{ShowMe}},
but with no option to include observations.
}
\examples{
 ShowMeDesign(strata.path="C:/Habitat.shp", transect.path="My2016Transects.shp")

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
