% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMeDouble}
\alias{ShowMeDouble}
\title{Display a \code{\link{DoubleObsRandom}} solution}
\usage{
ShowMeDouble(strata.path, solution)
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{solution}{The solution from a \code{\link{DoubleObsRandom}} optimization.}
}
\value{
None
}
\description{
ShowMeDouble will display the solution found by \code{\link{DoubleObsRandom}} on
the associated strata using Leaflet.
}
\details{
ShowMeDouble is designed to take a directory path for a stratification base layer
and the result of \code{\link{DoubleObsRandom}} and produce an interactive Leaflet map.
Clicking on strata will bring up the STRATNAME field associated with the click.
Clicking on a transect will display the associated strata identification, transect numbering,
recalculated (SPLIT) transect number (see \code{\link{SplitDesign}}), and
optimal positioning of the third observer (behind pilot or front-seat observer).
}
\examples{
 ShowMeDouble(strata.path="C:/Habitat.shp", solution=DoubleObsRandom(strata=MyStrata, transects=MyTransects))

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
