% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMeTrackDesign}
\alias{ShowMeTrackDesign}
\title{Display a combination of strata, transect design, and actual track flown}
\usage{
ShowMeTrackDesign(strata.path, transect.path, track.path, observer, years)
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.  Must be in a format
accepted by \code{\link{SplitDesign}}.}

\item{transect.path}{The path name to the .shp file of lines designating the transect design.
Must be in a format accepted by \code{\link{SplitDesign}}.}

\item{track.path}{The path name to the \emph{folder} containing the track files.
Track files must be .txt extension and contain \strong{lat, long, time, date.}}

\item{observer}{The observer initials to be used for track file selection.  Track files must
follow standard naming convention for track file objects to select observers correctly.}

\item{years}{The years of track files to display.}
}
\value{
None
}
\description{
ShowMeTrackDesign will take input layers of strata and transect design and display them using Leaflet with actual track flown overlay.
}
\details{
ShowMeTrackDesign is designed to take a directory path for a combination of stratification base layer,
aerial transect design, and actual track flown and display it using an interactive Leaflet map.
Clicking on strata will bring up the STRATNAME field associated with the click.
Clicking on a transect will display the
associated strata identification, transect numbering, recalculated (SPLIT) transect number
(see \code{\link{SplitDesign}}), and length in kilometers.  Similar to \code{\link{ShowMe}},
but with no option to include observations.
}
\examples{
 ShowMeTrackDesign(strata.path="C:/Habitat.shp", transect.path="My2016Transects.shp", track.path="c:/track files/", observer="CJF", years=2019)

}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
