% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMeUncut}
\alias{ShowMeUncut}
\title{Display a combination of unprocessed strata and transect design}
\usage{
ShowMeUncut(strata.path, transect.path, data.path = "none")
}
\arguments{
\item{strata.path}{The path name to the .shp file of the stratification.}

\item{transect.path}{The path name to the .shp file of lines designating the transect design.}

\item{data.path}{The path name(s) to the data file(s).  At minimum, data file must have columns
Observer, Species, Obs_Type, and Num.}
}
\value{
None
}
\description{
ShowMeUncut will take input layers of strata and transect design display them (unprocessed) using Leaflet.
}
\details{
ShowMeUncut is designed to take a directory path for a combination of stratification base layer,
aerial transect design and display them using an interactive Leaflet map.
Clicking on strata will bring up the STRATNAME field associated with the click.
Clicking on a transect will display the associated strata identification and transect numbering.
This is different than other ShowMe functions in that it uses unprocessed transect files that
haven't been applied and clipped to the associated strata layer.  Transects will still be displayed
if they fall outside strata boundaries.
}
\examples{
 ShowMeUncut(strata.path="C:/Habitat.shp", transect.path="My2016Transects.shp")

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
