% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ShowMeYears}
\alias{ShowMeYears}
\title{Display multiple years of observation data by survey area}
\usage{
ShowMeYears(area, year, species = "all")
}
\arguments{
\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta MBM duck stratification
\item YKDV - Yukon Kuskokwim Delta VCF study duck stratification
\item YKG - Yukon Kuskokwim Delta MBM goose stratification
\item KIG - Kigigak Island only
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
}}

\item{year}{The year or range of years to display.}

\item{species}{The species code(s) to be displayed.
Acceptable values are those in \code{sppntable}.}
}
\value{
None
}
\description{
ShowMeYears will take input data for an area and range of years and display them using Leaflet.
}
\details{
ShowMeYears is designed to take a combination of area and year(s) and look up associated files
in the \code{MasterFileList} to display the exact stratification and data that
would be used to generate index estimates using an interactive Leaflet map.
Data can be filtered to a single species or group of species. Clicking on strata will bring up
the STRATNAME field associated with the click. Clicking an observation will display
the observer initials, species code, observation type, number of the type observed,
transect number \emph{reported by the pilot or observer.}, and year of the observation.
}
\examples{
 ShowMeYears(area="ACP", year=c(2015:2019), species=c("SPEI", "STEI"))

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
