% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{ShowWBPHS}
\alias{ShowWBPHS}
\title{Display observations on WBPHS segments}
\usage{
ShowWBPHS(
  data,
  trans.file = "Q:/Waterfowl/WBPHS/Design Files/Design_Transects/NAWBPS_segments.shp"
)
}
\arguments{
\item{data}{WBPHS clean (greenlighted) observations.}

\item{trans.file}{Directory path to the WBPHS segment shape file (default to Q:/Waterfowl/WBPHS/Design Files/Design_Transects/NAWBPS_segments.shp)}
}
\value{
Interactive map of segment-level data
}
\description{
ShowWBPHS will load a leaflet display of WBPHS design segments and observations
}
\details{
ShowWBPHS provides a quick spatial overview of a set of observations and the Waterfowl Breeding Population Habitat Survey segments.
I am unsure of the origin and accuracy of the segments.  Segments and observations are drawn using leaflet.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
