% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SpeciesByProject}
\alias{SpeciesByProject}
\title{Changes species codes based on area}
\usage{
SpeciesByProject(full.data, area)
}
\arguments{
\item{full.data}{The data frame containing the observations}

\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKG - Yukon Kuskokwim Delta, geese
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}
}
\value{
data frame with area-specific species codes
}
\description{
SpeciesByProject will look up species codes and potentially alter them based on project area
}
\details{
SpeciesByProject is designed to alter species codes by area specifications. Current lists can be found in \code{sppntable}.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
