% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{SpeciesCheck}
\alias{SpeciesCheck}
\title{Check if Species column inputs are all correct}
\usage{
SpeciesCheck(data, area)
}
\arguments{
\item{col.data}{The column to be checked.}
}
\value{
data frame of TRUE/FALSE, incorrect unfixable entries, and incorrect but fixable entries
}
\description{
SpeciesCheck checks to see if Species entries are correct values needed for analysis.
}
\details{
SpeciesCheck is a GreenLight QA/QC check performed on a data set.  Species entries are cross-referenced with approved entries by project in
\code{\link{sppntable}} and values with known treatments are marked as "yellow light" issues that will be updated.  Completely unknown species codes
will result in a "red" status.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
