% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SplitDesign}
\alias{SplitDesign}
\title{Renumber design transects based on stratification}
\usage{
SplitDesign(strata.file, transect.file, SegCheck = FALSE, area = "other")
}
\arguments{
\item{strata.file}{The location of the analysis stratification .shp file}

\item{transect.file}{The location of the design transect .shp file (currently only supports east-west transects)}

\item{SegCheck}{Should a map be drawn to verify the process?  TRUE/FALSE}

\item{area}{If area=="CRD" a specific fix is made to EggIsland north/south transects.  Otherwise area does not matter.}
}
\value{
spatial lines object with appropriate numbering (retains old numbering)
}
\description{
SplitDesign will read transect and stratification layers and number transects based on polygon arrangement in the stratification file
}
\details{
SplitDesign is a critically important function to aerial survey analysis.  It will read in design transect and design stratifcation and combine
the polygons and lines to renumber the design file appropriately.  It will catch errors such as slight overhang in design transects and strata,
incorrect numbering scheme for a given sample, and incorrectly attributed strata to lines.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
