% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{SplitWBPHS}
\alias{SplitWBPHS}
\title{Combine and parse raw WBPHS data files}
\usage{
SplitWBPHS(folder.path)
}
\arguments{
\item{folder.path}{The directory path to the folder for input and output files.}
}
\value{
Year-Observer combinations of data in csv format to folder.path
}
\description{
SplitWBPHS will combine multiple observation files and split into annual observer files
}
\details{
SplitWBPHS takes a folder of raw text files (generally at the transect or stratum level) and combines them into an annual data file
for an observer using the naming convention: WBPHS_YYYY_RawObs_NNN.csv, where YYYY is the 4 digit year and NNN is 3 character initials.
Columns must be in the following order: Year, Month, Day, Seat, Observer, Stratum, Transect, Segment, Flight_Dir, A_G_Name, Wind_Dir,
Wind_Vel, Sky, Filename, Lat, Lon, Time, Delay, Species, Num, Obs_Type.
It will also add the columns Behavior, Distance, Code, and Notes that aren't used in the standard WBPHS protocol, but have become
the standard in Alaska Region aerial surveys.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
