% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSpace.R
\name{StateSpace}
\alias{StateSpace}
\title{Runs and reports on a generic state space model of population growth}
\usage{
StateSpace(
  area,
  index,
  years,
  species,
  N1 = c(1000, 1e+05),
  r = c(-0.3, 0.3),
  sigma = c(0, 0.3),
  n.chains = 3,
  n.thin = 1,
  n.iter = 5000,
  n.burnin = 1000
)
}
\arguments{
\item{area}{The area code for dedicated MBM Alaska region surveys.  Either folder.path or area must be specified.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKG - Yukon Kuskokwim Delta, geese
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
}}

\item{index}{The index used.
 Acceptable values include:
\itemize{
\item itotal - indicated total; singles and pairs doubled, flocks added, flkdrake added, flkdrake 2-4 doubled
\item ibb - indicated breeding birds; singles and pairs doubled
\item total - total birds observed; pairs doubled, others added}}

\item{years}{- The years to run the model over}

\item{species}{- The accepted \code{sppntable} species code for which to model}

\item{N1}{- Initial population size in the format c(min, max) to establish log-scale uniformly-distributed prior}

\item{r}{- Range of growth rate parameter in the format c(min, max) to establish normally-distributed prior}

\item{sigma}{- Range of the process standard deviation parameter in the format c(min, max) to establish uniformly-distributed prior}

\item{n.chains}{- Number of Monte Carlo chains}

\item{n.thin}{- Model thin rate}

\item{n.iter}{- Model iterations}

\item{n.burnin}{- Burn-in iteration length}
}
\value{
Will generate an .html report to the current working directory
}
\description{
StateSpace will search the internal AKaerial index estimates and run a generic state space model over the range selected
}
\details{
A state-space model is a general term that usually refers to a model containing the true underlying (and unobserved)
 time-dependent state of the system and imperfect time-dependent observations.
 The states change with time so that the state at time t depends on the state at time t-1
 (and potentially other factors).  Here, the state is defined as the unobserved total population of geese or ducks in the survey area, and observations depend just on the state.
 A formal description is contained in GenericStateSpaceModel.Rmd, which is provided as the output to this function
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
\itemize{
\item{Charles Frost, \email{charles_frost@fws.gov}}
\item{Erik Osnas, \email{erik_osnas@fws.gov}}}
}
