% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{StrataSummary}
\alias{StrataSummary}
\title{Summarize design strata for analysis}
\usage{
StrataSummary(strata.file)
}
\arguments{
\item{strata.file}{The path to the design stratification .shp file}
}
\value{
data frame summary of stratification
}
\description{
StrataSummary will provide the overall study area spatial characteristics for analysis
}
\details{
Stratasummary will compute areas of each strata in a design strata file as well as calculate the maximum possible transects in
a sample (M) using FindVoids.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
