% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataSummarySF.R
\name{StrataSummarySF}
\alias{StrataSummarySF}
\title{Summarize design strata for analysis using sf package}
\usage{
StrataSummarySF(strata.file, id = "STRATNAME")
}
\arguments{
\item{strata.file}{The path to the design stratification file}

\item{id}{The identifier to group strata by}
}
\value{
data frame summary of stratification
}
\description{
StrataSummarySF will provide the overall study area spatial characteristics for analysis
}
\details{
StratasummarySF will compute areas of each strata (in km^2) in a design strata file as well as calculate the maximum possible transects in
a sample (M).
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
