% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{SummaryWBPHS}
\alias{SummaryWBPHS}
\title{Summarize the flight information for an observer on the WBPHS}
\usage{
SummaryWBPHS(full.data, strip.width = 0.25)
}
\arguments{
\item{full.data}{A clean (greenlight) file of observations by stratum, transect, and segment}

\item{strip.width}{The sampled strip width in miles}
}
\value{
Data frame containing strata, transect, and segment information by observer
}
\description{
SummaryWBPHS will summarize the flight and sampled area of a pilot or observer and give both the original and renumbered transect for reference
}
\details{
This function takes an observation file and uses it to imply the sampled area for each observer.  Segment, transect, and strata records are pulled
from the observations and compared to a table of known values for lengths and sizes.  Segments in all strata except stratum 7 and stratum 12
are 16 miles long.  Segments in stratum 7 are 8 miles long and segments in stratum 12 are 18 miles long.  Total lengths for the augmented transects 12-19
in modified stratum 9 (see Cut9) are:
\itemize{
  \item 12:  22 mi
  \item 13:  20.763 mi
  \item 14:  21.32 mi
  \item 15:  12 mi
  \item 16:  14.543 mi
  \item 17:  8.395 mi
  \item 18:  7.663 mi
  \item 19:  8.372 mi
  }
 Default strip width is 0.25 mi for 2 observers, each viewing 0.125 mi.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
