% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{SwanCheck}
\alias{SwanCheck}
\title{Check if swan and swan nest entries are all correct}
\usage{
SwanCheck(data)
}
\arguments{
\item{data}{The data frame to be checked.}
}
\value{
data frame of TRUE/FALSE and offending entries
}
\description{
SwanCheck checks to see if swan and swan nest entries were transcribed correctly for analysis.
}
\details{
SwanCheck is a GreenLight QA/QC check performed on a data set.  The correct transcription of a swan and associated nest is to
record the swan Obs_Type (single or pair) and then the associated nest as a separate row with Obs_Type open and Num 1.  Anything
else will result in "yellow" status if the treatment of the data type is known, and "red" if unknown.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
