% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{SwitchMatch}
\alias{SwitchMatch}
\title{Check for the old style of transcribed data.}
\usage{
SwitchMatch(data)
}
\arguments{
\item{data}{The data frame to be checked.}
}
\value{
TRUE/FALSE
}
\description{
SwitchMatch will test if the original style of transcribed column headings is present
}
\details{
SwitchMatch is designed to check if the old style of column headings is present and the minimum set of
yr, se, mo, strat, da, grp, and sppn exists.  If TRUE, then ColSwitch can be used to overwrite to the current headings.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
