% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyDensities.R
\name{TidyDensities}
\alias{TidyDensities}
\title{Tidy standard ratio estimator for aerial survey data}
\usage{
TidyDensities(data, area, output = FALSE)
}
\arguments{
\item{data}{The DataProcess list object to be analyzed}

\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
\item YKG - Yukon Kuskokwim Delta, geese
\item KIG - Kigigak Island
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}

\item{output}{TRUE/FALSE Should results be written to text files in the current working directory?}
}
\value{
List object with 2 elements: \enumerate{
\item estimates Observer-specific estimates for the species indicated in the sppntable estimates column
\item counts.final Deeper level count information by transect, strata, species, and observer
}
}
\description{
TidyDensities will combine spatially-referenced observations with design transects and strata to create an index estimate
}
\details{
TidyDensities is the primary function in AKaerial for producing index estimates. It will take an object from DataSelect and adjust counts, summarize
spatial information, and calculate indices and their associated standard errors.  Also retained are estimates of densities of bird by strata and on
 each transect.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
