% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TransData}
\alias{TransData}
\title{Aggregate observations by Obs_Type, corrected transect, species, observer, and year}
\usage{
TransData(selected.data)
}
\arguments{
\item{selected.data}{The data object from DataSelect}
}
\value{
data frame with transect-level summary of observations
}
\description{
TransData will provide an aggregated transect-level summary of all observations in a DataSelect output list by observer and year
}
\details{
TransData is designed to take an object from DataSelect and summarize observations at the transect level.  Each row is a unique
combination of year, observer, species, transect, and Obs_Type.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
