% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransDataSF.R
\name{TransDataSF}
\alias{TransDataSF}
\title{Aggregate observations by Obs_Type, corrected transect, species, observer, and year}
\usage{
TransDataSF(selected.data, strata.id)
}
\arguments{
\item{selected.data}{The data object from DataSelect}
}
\value{
data frame with transect-level summary of observations
}
\description{
TransDataSF will provide an aggregated transect-level summary of all observations in a DataSelect output list by observer and year
}
\details{
TransDataSF is designed to take an object from DataSelect and summarize observations at the transect level.  Each row is a unique
combination of year, observer, species, transect, and Obs_Type.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
