% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{TransDataWBPHS}
\alias{TransDataWBPHS}
\title{Summarize index counts for WBPHS data by Segment, Transect, and Stratum}
\usage{
TransDataWBPHS(selected.data)
}
\arguments{
\item{selected.data}{An R data frame of observations in a given year of the survey}
}
\value{
A data frame with index values by Year, Observer, Species, Obs_Type, Segment, Transect, and Stratum.
}
\description{
Summarize index counts for WBPHS data by Segment, Transect, and Stratum
}
\details{
This function takes standard greenlight data for the WBPHS (Waterfowl Breeding Population Habitat Survey, or North American) and
summarizes it at the Segment, Transect, and Stratum level.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
