% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransectFile.R
\name{TransectFile}
\alias{TransectFile}
\title{Create or append to a csv file with study area transect panel information}
\usage{
TransectFile(
  method = "create",
  output.folder = "none",
  append.to = NULL,
  append.year = NULL,
  append.area = NULL
)
}
\arguments{
\item{method}{"create" or "append"}

\item{append.to}{File path for appending}
}
\value{
Text file (.csv) is opened, appended to, and saved, or created and saved.
}
\description{
TransectFile will summarize study area transect-level information into a csv file
}
\details{
TransectFile will take a spatial layer with study area stratification polygons and
transect line files and summarize the lengths into a csv file by year.  The file can either be
created entirely or appended to for a single year addition.
The file is intended to be an input into the estimate-generating functions in AKaerial.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
