% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TransSummary}
\alias{TransSummary}
\title{Summarize the flight information for an observer}
\usage{
TransSummary(full.data, split.design, area)
}
\arguments{
\item{full.data}{A clean (greenlight) file of observations}

\item{split.design}{SplitDesign output object of corrected transect lines}

\item{area}{The area code for dedicated MBM Alaska region surveys.
  Acceptable values include:
\itemize{
\item YKD - Yukon Kuskokwim Delta, ducks
\item YKDV - Yukon Kuskokwim Delta, 2018 visibility study
\item YKG - Yukon Kuskokwim Delta, geese
\item KIG - Kigigak Island
\item ACP - Arctic Coastal Plain
\item CRD - Copper River Delta
\item WBPHS - The North American Waterfowl Breeding Population Habitat Survey
\item BLSC - 2018 Black Scoter Survey
}}
}
\value{
Transect summary data frame
}
\description{
TransSummary will summarize the flight and sampled area of a pilot or observer and give both the original and renumbered transect for reference
}
\details{
TransSummary uses what was declared by a pilot or observed as navigational transect number in their transcribed data file and cross references it
with the corrected transect number (ctran) in split design to develop a flight record for the pilot/observer.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
