% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TrimToStrata}
\alias{TrimToStrata}
\title{Trim observations that fall outside of the area of inference}
\usage{
TrimToStrata(full.data, strata.path)
}
\arguments{
\item{full.data}{A clean (greenlight) file of observations}

\item{strata.path}{The path to the design stratification .shp file}
}
\value{
data frame of full.data without egregious observations
}
\description{
TrimToStrata will clip observation data to the strata polygons
}
\details{
TrimToStrata is especially important on the Copper River Delta (CRD) study area, where design transects intentionally sample outside
of the area of inference.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
