% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlight.R
\name{UnitCheck}
\alias{UnitCheck}
\title{Check if Obs_Type entries are all correct}
\usage{
UnitCheck(data)
}
\arguments{
\item{col.data}{The column to be checked.}
}
\value{
data frame of TRUE/FALSE and offending entries
}
\description{
UnitCheck checks to see if Obs_Type entries are correct values needed for analysis.
}
\details{
UnitCheck is a GreenLight QA/QC check performed on a data set.  If the Obs_Type entries are not single, pair, flkdrake, or open, it will return a "red" status.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
