% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenlightscribe.R
\name{UnitCheckScribe}
\alias{UnitCheckScribe}
\title{Check if Grouping entries are all correct}
\usage{
UnitCheckScribe(data)
}
\arguments{
\item{col.data}{The column to be checked.}
}
\value{
data frame of TRUE/FALSE and offending entries
}
\description{
UnitCheckScribe checks to see if Grouping entries are correct values needed for analysis.
}
\details{
UnitCheckScribe is a GreenLight QA/QC check performed on a data set.  If the Grouping entries are not single, pair, flkdrake, or open, it will return a "red" status.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
