% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ViewStrata}
\alias{ViewStrata}
\title{Display stratification in plot format}
\usage{
ViewStrata(
  map,
  year = NULL,
  ViewTrans = "none",
  strata = "all",
  numbers = FALSE,
  print = TRUE
)
}
\arguments{
\item{map}{The LoadMap object to be displayed}

\item{year}{The year of the desired transect panel}

\item{ViewTrans}{Should the transects be plotted?  TRUE/FALSE}

\item{strata}{Character string of strata to be plotted (or "all")}

\item{numbers}{Should transects be numbered?  TRUE/FALSE}

\item{print}{Should the result be plotted (TRUE) or saved (FALSE)}
}
\value{
ggplot2 plot object
}
\description{
ViewStrata uses ggplot2 to display a stratification map.
}
\details{
This function was replaced with a series of ShowMe functions and is no longer used.  See \code{ShowMe}.
}
\references{
\url{https://github.com/cfrost3/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
