% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBPHS_VCF-doc.R
\docType{data}
\name{WBPHS_VCF}
\alias{WBPHS_VCF}
\title{Current table of VCFs (visibility correction factors) for relevant species on the WBPHS survey}
\format{

}
\usage{
data(WBPHS_VCF)
}
\description{
Current table of all species on the USFWS Region 11 Migratory Bird Management component of the North American (WBPHS) aerial survey
and their associated visibility correction factors (VCFs).
}
\keyword{datasets}
