% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{WBPHSbyYear}
\alias{WBPHSbyYear}
\title{Calculate index estimates for a given year of WBPHS count and transect data}
\usage{
WBPHSbyYear(year)
}
\arguments{
\item{year}{The 4-digit year requested to estimate generation}
}
\value{
Data frame of counts, densities, areas, index estimates, and adjusted index estimates.
}
\description{
WBPHSbyYear pulls annual WBPHS data specified by MasterFileList_WBPHS to calculate an index estimate using a ratio estimator
}
\details{
WBPHSbyYear provides a wrapper for WBPHStidy that will select the appropriate data files for generating an annual index estimate for the WBPHS survey.
MasterFileList_WBPHS contains a list of "official" files and folder paths for data used in WBPHS estimates tables.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
