% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{WBPHSMultipleYear}
\alias{WBPHSMultipleYear}
\title{Calculate index estimates for a given year of WBPHS count and transect data}
\usage{
WBPHSMultipleYear(years)
}
\arguments{
\item{years}{The range of years requested to estimate generation, currently available for 1957-2021.}
}
\value{
Data frame of counts, densities, areas, index estimates, and adjusted index estimates.
}
\description{
WBPHSMultipleYear iterates over a range of years to produce multiple years of index estimates using a ratio estimator
}
\details{
WBPHSMultipleYear provides a wrapper for WBPHSbyYear that will select the appropriate data files for generating annual index estimates for the WBPHS survey.
MasterFileList_WBPHS contains a list of "official" files and folder paths for data used in WBPHS estimates tables. The final table includes year- and species-specific
deletions based on annual variation in, changes to, and interpretations of the standardized protocol.  The unavailable estimates include: \enumerate{
\item BRAN, EMGO, SACR, SNGO, SWAN, SWANN from 1957 to 1963
\item COLO, PALO, RTLO, UNLO, YBLO from 1957 to 1970
\item HOGR, RNGR, UNGR from 1957 to 1990
\item BAEA from 1957 to 1964
\item GOEA from 1957 to 2016
\item CCGO, GWFG from 1957 to 1963, other than total estimates (augmented indices unavailable)
}
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
