% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthAmerican.R
\name{WBPHStidy}
\alias{WBPHStidy}
\title{Calculate index estimates for specified WBPHS count and transect data}
\usage{
WBPHStidy(adj, flight)
}
\arguments{
\item{adj}{Adjusted counts for a given year of the WBPHS survey (see ReadWBPHS)}

\item{flight}{Flight summary for a given year of the WBPHS survey (see SummaryWBPHS)}
}
\value{
Data frame of counts, densities, areas, index estimates, and adjusted index estimates.
}
\description{
WBPHStidy uses annual WBPHS adjusted count data and flown transects to calculate an index estimate using a ratio estimator
}
\details{
WBPHStidy provides a method for creating an annual index estimate for the WBPHS survey.  Individual observations and sampled areas are summed first at the transect level.
Groups are created for eiders, mergansers, scoters, and grebes that pool species as: \enumerate{
\item Eider COEI, KIEI, SPEI, STEI, UNEI
\item Merganser COME, RBME, UNME
\item Scoter BLSC, SCOT, SUSC, WWSC
\item Grebe HOGR, RNGR, UNGR
}
Transect-level counts are summarized at the stratum level to produce densities of index groupings and variance according to the classic ratio estimator (Cochran 1977),
then index estimates are adjusted according to historical visual correction factors (VCFs, see WBPHS_VCF) in an attempt to correct for incomplete detection.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
