% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKDVEst.R
\name{YKDVEst}
\alias{YKDVEst}
\title{YKDVEst uses AKaerial tables to generate a VCF-corrected estimate of SPEI on the YKD}
\usage{
YKDVEst()
}
\value{
A list containing population estimates of 'indicated breeding birds', the 'index" estimates (without detection correction), standard errors of each, and a ggplot objects of each by year.
}
\description{
YKDVEst uses AKaerial tables to generate a VCF-corrected estimate of SPEI on the YKD
}
