% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsFile.R
\name{ObsFile}
\alias{ObsFile}
\title{Create or append to a master csv file of observations}
\usage{
ObsFile(method = "create", append.to = NULL)
}
\arguments{
\item{method}{"create" or "append"}

\item{append.to}{File path for appending}
}
\value{
Text file (.csv) is opened, appended to, and saved, or created and saved.
}
\description{
ObsFile will summarize all observation information into a csv file
}
\details{
ObsFile will take a spatial layer with study area stratification polygons and
transect lengths and summarize the observations into a csv file.  The file can either be created entirely or appended
to for a single year addition.  The file is intended to be an input into the estimate-generating
functions in AKaerial.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
