% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ReportFigure}
\alias{ReportFigure}
\title{Create a basic figure from historic estimates, including n-year average}
\usage{
ReportFigure(
  data,
  species,
  year,
  index = "none",
  yr.avg = 0,
  title = "none",
  x.label = "none",
  y.label = "none",
  leg.values = 0,
  leg.labels = "none",
  leg.limits = "none",
  test.out = TRUE
)
}
\arguments{
\item{data}{The estimates object.  Must be in a format with Year, Species, index, and index variance if not one of
the included package data objects (ACPHistoric, CRDHistoric, YKDHistoric, YKDVHistoric, or YKGHistoric).}

\item{species}{The species chosen (see \code{\link{sppntable}} for options).  Currently
supports no more than 1 species.}

\item{year}{The range of years for the figure.}

\item{index}{The column names that specify the index estimate and its variance.
Must be ordered correctly (index, variance) or (index1, variance1, index2, variance2).  Currently up to 2 indices are supported if
a multi-year average is requested, or 4 indices if no multi-year average is requested.
If no index is specified, the user will be prompted with choices from the data.}

\item{yr.avg}{The number of years in the n-year running average requested.}

\item{title}{The overall figure title.  If nothing entered as an argument, the user will be prompted to enter a title in the function.}

\item{x.label}{The desired x axis label for the final figure.  If this is not specified the user will be prompted to
input within the function.}

\item{y.label}{The desired y axis label for the final figure.  If this is not specified the user will be prompted to
input within the function.}

\item{leg.values}{The desired reordering of the default legend order.  The user should enter this as a series of numbered positions in the current
figure legend.  For example, reversing a 4-element legend would be entered 4,3,2,1.  Each number separated only by a comma.}

\item{leg.labels}{The desired legend labels in the new desired order.  These should be entered as text and separated only by a comma.}

\item{leg.limits}{The internal data object references in the new order.  This specifies the mapping of the newly-defined labels and orders.
The easiest way to get these is to look at the FIRST figure generated by the function and enter them as they appear in the legend, only in the
new order.  They MUST match the original figure names/column names.}

\item{test.out}{Should the figure progression output in the plot window?  If FALSE, the object is only returned.}
}
\value{
Renders and returns a ggplot figure
}
\description{
ReportFigure will summarize historic estimates from Alaska Region aerial surveys in a line graph that can be exported for reporting.
}
\details{
ReportFigure will take one of the historic estimates tables (package data for AKaerial) and display it as a ggplot figure
suitable for an external report.  The user can specify n for an n-year average if one is desired.  A similarly
structured data frame could be used in place of one of the historic estimates tables, provided it has columns for Year,
Species, an index, and the index variance.  Figure title and legend text can either be specified as arguments or
entered by the user as the figure is generated.
}
\examples{
ReportFigure(data=YKGHistoric$combined,
species="EMGO",
index=c("total", "total.var", "itotal", "itotal.var", "ibb", "ibb.var"),
year=c(2007:2019),
title="Emperor goose population indices from the Yukon-Kuskokwim Delta, 2007-2019",
x.label = "Year",
y.label="Index Estimate",
leg.values = c(3,2,1),
leg.labels = c("Total", "Indicated Total", "Indicated Breeding"),
leg.limits = c("total", "itotal", "ibb"))

}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
