% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataFile.R
\name{StrataFile}
\alias{StrataFile}
\title{Create or append to a csv file with study area stratum information}
\usage{
StrataFile(method = "create", append.to = NULL)
}
\arguments{
\item{method}{"create" or "append"}

\item{append.to}{File path for appending}
}
\value{
Text file (.csv) is opened, appended to, and saved, or created and saved.
}
\description{
StrataFile will summarize study area stratum-level information into a csv file
}
\details{
StrataFile will take a spatial layer with study area stratification polygons and
summarize the areas into a csv file.  The file can either be created entirely or appended
to for a single year addition.  The file is intended to be an input into the estimate-generating
functions in AKaerial.
}
\references{
\url{https://github.com/USFWS/AKaerial}
}
\author{
Charles Frost, \email{charles_frost@fws.gov}
}
